requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
langchain==0.3.3
langgraph==0.2.35
requests==2.32.3
python-dotenv==1.0.1